Fix remove investment from sidebar
When trying to remove an investment from the sidebar, we are getting a response status of (blocked:mixed-content).

This is due to the budget_ballot_line_url being http instead of https. By using path, instead of url, the link is generated with https.

There is a deeper issue here, url should also use https. It could be fixed by reviewin the nginx configuration for ssl or other rails configuration parameters.